Add SPICE submodule
